requirements.txt
# FDB Log Analyzer Dependencies
# Python 3.8+

# 核心依赖
duckdb>=0.9.0
pydantic>=2.0.0

# 可选依赖（用于数据导出）
pandas>=2.0.0
pyarrow>=14.0.0  # 用于parquet格式导出

# AI Agent依赖 (核心组件)
google-genai  # Google Gemini API - AI Agent uses Gemini 2.5 Flash for analysis
